Update engine example
